pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastapi-production-starter"
version = "1.0.0"
description = "Production-ready FastAPI starter template with security, observability, and modern Python best practices built-in."
readme = "README.md"
authors = [{ name = "Siddhant Shah" }]
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["fastapi", "starter", "template", "production", "api", "async", "security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
  "fastapi>=0.111,<0.112",
  "uvicorn[standard]>=0.30,<0.31",
  "pydantic-settings>=2.3,<3.0",
  "structlog>=24.2,<25.0",
  "orjson>=3.10,<4.0",
  "python-dotenv>=1.0,<2.0"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.4,<0.5",
  "black>=24.4,<25.0",
  "mypy>=1.10,<2.0",
  "pytest>=8.2,<9.0",
  "pytest-asyncio>=0.23,<0.24",
  "httpx>=0.27,<0.28",
  "pytest-cov>=5.0,<6.0"
]

# Phase 2: Redis and JWT authentication
phase2 = [
  "redis>=5.0,<6.0",
  "hiredis>=2.2,<3.0",
  "pyjwt>=2.8,<3.0",
  "cryptography>=41.0,<42.0"
]

# Phase 3: Advanced features
phase3 = [
  "geoip2>=4.7,<5.0",
  "prometheus-client>=0.19,<0.20"
]

# All security features combined
security = [
  "redis>=5.0,<6.0",
  "hiredis>=2.2,<3.0",
  "pyjwt>=2.8,<3.0",
  "cryptography>=41.0,<42.0",
  "geoip2>=4.7,<5.0",
  "prometheus-client>=0.19,<0.20"
]

[project.urls]
homepage = "https://github.com/thesidshah/fastapi-production-starter"
repository = "https://github.com/thesidshah/fastapi-production-starter"
documentation = "https://github.com/thesidshah/fastapi-production-starter#readme"
issues = "https://github.com/thesidshah/fastapi-production-starter/issues"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "--strict-config --strict-markers"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "UP",
  "B",
  "A",
  "COM",
  "ISC",
  "C4",
  "ARG"
]
ignore = ["B008"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
disallow_any_generics = true
warn_unused_ignores = true
show_error_codes = true
pretty = true
namespace_packages = true
explicit_package_bases = true
mypy_path = "src"

